Remove duplicate Convert.hpp header and add doxygen
